refactor(18025): add dynamic/lazy import - reduce svg files size
